requirements.txt
pygubu
pycrypto
pillow
